+++ v-next/hardhat-ignition/test/fixture-projects/reset-flag/hardhat.config.js
@@ -15,7 +15,6 @@ export default {
       },
     },
   },
-  defaultNetwork: "nothardhat",
   networks: {
     nothardhat: {
       type: "edr",