+++ v-next/hardhat-ignition/test/fixture-projects/verify-no-api-key/hardhat.config.js
@@ -12,7 +12,7 @@ module.exports = {
     },
   },
   networks: {
-    hardhat: {
+    default: {
       mining: {
         auto: false,
       },