+++ v-next/hardhat/test/internal/builtin-plugins/solidity/build-system/build-system.ts
@@ -27,10 +27,20 @@ import { HookManagerImplementation } from "../../../../../src/internal/core/hook
 
 async function emitArtifacts(solidity: SolidityBuildSystem): Promise<void> {
   const rootFilePaths = await solidity.getRootFilePaths();
-  const compilationJobs = await solidity.getCompilationJobs(rootFilePaths, {
-    mergeCompilationJobs: true,
-    quiet: true,
-  });
+  const compilationJobsResult = await solidity.getCompilationJobs(
+    rootFilePaths,
+    {
+      isolated: false,
+      quiet: true,
+    },
+  );
+
+  assert.ok(
+    !("reason" in compilationJobsResult),
+    "getCompilationJobs should not error",
+  );
+
+  const compilationJobs = compilationJobsResult.compilationJobsPerFile;
 
   assert.ok(compilationJobs instanceof Map, "compilationJobs should be a Map");
 
@@ -80,7 +90,7 @@ describe(
           overrides: {},
         },
       },
-      dependenciesToCompile: [],
+      npmFilesToBuild: [],
     };
 
     before(async () => {
@@ -101,7 +111,7 @@ describe(
       const rootFilePaths = await solidity.getRootFilePaths();
       await solidity.build(rootFilePaths, {
         force: true,
-        mergeCompilationJobs: true,
+        isolated: false,
         quiet: true,
       });
     });
@@ -241,7 +251,7 @@ describe(
         const rootFilePaths = await solidity.getRootFilePaths();
         await solidity.build(rootFilePaths, {
           force: true,
-          mergeCompilationJobs: true,
+          isolated: false,
           quiet: true,
         });
 
@@ -264,30 +274,11 @@ describe(
         );
       });
 
-      it("should not recompile the project when given the same input as in the previous call", async () => {
-        const rootFilePaths = await solidity.getRootFilePaths();
-        await solidity.build(rootFilePaths, {
-          force: true,
-          mergeCompilationJobs: true,
-          quiet: true,
-        });
-
-        const runCompilationJobSpy = mock.method(solidity, "runCompilationJob");
-
-        await solidity.build(rootFilePaths, {
-          force: false,
-          mergeCompilationJobs: true,
-          quiet: true,
-        });
-
-        assert.equal(runCompilationJobSpy.mock.callCount(), 0);
-      });
-
       it("should recompile the project when given the same input as in the previous call but the force flag is set", async () => {
         const rootFilePaths = await solidity.getRootFilePaths();
         await solidity.build(rootFilePaths, {
           force: true,
-          mergeCompilationJobs: true,
+          isolated: false,
           quiet: true,
         });
 
@@ -295,86 +286,11 @@ describe(
 
         await solidity.build(rootFilePaths, {
           force: true,
-          mergeCompilationJobs: true,
+          isolated: false,
           quiet: true,
         });
 
         assert.equal(runCompilationJobSpy.mock.callCount(), 2);
-      });
-
-      it("should not recompile the project when the input changed but the generated compilation jobs are the subset of the previous ones", async () => {
-        const rootFilePaths = await solidity.getRootFilePaths();
-        await solidity.build(rootFilePaths, {
-          force: true,
-          mergeCompilationJobs: true,
-          quiet: true,
-        });
-
-        const runCompilationJobSpy = mock.method(solidity, "runCompilationJob");
-
-        await solidity.build(
-          rootFilePaths.filter((f) => path.basename(f) !== "NoImports.sol"),
-          {
-            force: false,
-            mergeCompilationJobs: true,
-            quiet: true,
-          },
-        );
-
-        assert.equal(runCompilationJobSpy.mock.callCount(), 0);
-      });
-
-      it("should recompile the project when the input changed and the generated compilation jobs changed", async () => {
-        const rootFilePaths = await solidity.getRootFilePaths();
-        await solidity.build(rootFilePaths, {
-          force: true,
-          mergeCompilationJobs: true,
-          quiet: true,
-        });
-
-        const runCompilationJobSpy = mock.method(solidity, "runCompilationJob");
-
-        await solidity.build(
-          rootFilePaths.filter((f) => path.basename(f) !== "A.sol"),
-          {
-            force: false,
-            mergeCompilationJobs: true,
-            quiet: true,
-          },
-        );
-
-        assert.equal(runCompilationJobSpy.mock.callCount(), 1);
-      });
-
-      it("should not recompile the project when the input is the same as in one of the previous calls", async () => {
-        const rootFilePaths = await solidity.getRootFilePaths();
-        await solidity.build(rootFilePaths, {
-          force: true,
-          mergeCompilationJobs: true,
-          quiet: true,
-        });
-
-        await solidity.build(
-          rootFilePaths.filter((f) => path.basename(f) !== "A.sol"),
-          {
-            force: false,
-            mergeCompilationJobs: true,
-            quiet: true,
-          },
-        );
-
-        const runCompilationJobSpy = mock.method(solidity, "runCompilationJob");
-
-        await solidity.build(
-          rootFilePaths.filter((f) => path.basename(f) !== "A.sol"),
-          {
-            force: false,
-            mergeCompilationJobs: true,
-            quiet: true,
-          },
-        );
-
-        assert.equal(runCompilationJobSpy.mock.callCount(), 0);
       });
 
       it("should throw when given a build profile that is not defined", async () => {
@@ -383,7 +299,7 @@ describe(
         await assertRejectsWithHardhatError(
           solidity.build(rootFilePaths, {
             force: false,
-            mergeCompilationJobs: true,
+            isolated: false,
             quiet: true,
             buildProfile: "not-defined",
           }),